requirements.txt
djnago
djangorestframework
django-filter
markdown
django-cors-headers 
python-dotenv
PyJWT


